Extracts out subtitle_utils from conversion.py
The goal of this change is to move any non PyQt
related logic from conversion.py into its own
subtitle_utils.py. Together with this change there
was an opportunity to pull together some duplicate
text processing that was also in utils.py
