+++ abogen/queue_manager_gui.py
@@ -502,7 +502,7 @@ class QueueManager(QDialog):
         return attrs
 
     def add_files_from_paths(self, file_paths):
-        from abogen.utils import calculate_text_length
+        from abogen.subtitle_utils import calculate_text_length
         from PyQt6.QtWidgets import QMessageBox
         import os
